# Conflicts:
#	FertiFront/src/app/servicios/fuente.service.ts
#	FertiWeb/bin/main/dataH2.sql
#	FertiWeb/src/main/resources/dataH2.sql

+++ FertiWeb/src/main/java/controladores/ControladorDatosTipoCultivo.java
@@ -10,8 +10,14 @@ import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.transaction.annotation.Transactional;
 
+import constantes.MensajesConstantes;
 import dominio.TipoCultivo;
+import excepciones.ExcepcionClaveForanea;
+import persistencia.entidad.ParcelaEntidad;
+import persistencia.entidad.RequerimientoEntidad;
 import persistencia.entidad.TipoCultivoEntidad;
+import persistencia.repositorio.ParcelaRepository;
+import persistencia.repositorio.RequerimientoRepository;
 import persistencia.repositorio.TipoCultivoRepository;
 
 public class ControladorDatosTipoCultivo extends ControladorDatos {
@@ -21,6 +27,12 @@ public class ControladorDatosTipoCultivo extends ControladorDatos {
 
 	@Autowired
 	private TipoCultivoRepository tipoCultivoRepository;
+
+	@Autowired
+	private RequerimientoRepository requerimientoRepository;
+
+	@Autowired
+	private ParcelaRepository parcelaRepository;
 
 	@Cacheable("GlobalCacheConstant.CACHE_CULTIVOS")
 	public List<TipoCultivo> consultarCultivo() {
@@ -37,7 +49,7 @@ public class ControladorDatosTipoCultivo extends ControladorDatos {
 	public TipoCultivo consultarTipoCultivoXId(Long codigoCultivoSembrado) {
 		return mapearADominio(tipoCultivoRepository.findByCodigoTipoCultivo(codigoCultivoSembrado));
 	}
-	
+
 	private List<TipoCultivo> mapearListaADominio(List<TipoCultivoEntidad> tipoCultivoEntidadList) {
 		return tipoCultivoEntidadList.stream().map(a -> mapearADominio(a))
 				.collect(Collectors.toCollection(ArrayList::new));
@@ -47,7 +59,7 @@ public class ControladorDatosTipoCultivo extends ControladorDatos {
 	protected Object construirDTO(Object object) {
 		throw new UnsupportedOperationException();
 	}
-	
+
 	@Override
 	protected Object construirDominio(Object object) {
 		throw new UnsupportedOperationException();
@@ -67,19 +79,26 @@ public class ControladorDatosTipoCultivo extends ControladorDatos {
 
 	@Transactional
 	@Override
-	@CacheEvict(value= {"GlobalCacheConstant.CACHE_CULTIVOS",
-			"GlobalCacheConstant.CACHE_CULTIVOS_APP", 
-			"GlobalCacheConstant.CACHE_CULTIVO"}, allEntries=true)
+	@CacheEvict(value = { "GlobalCacheConstant.CACHE_CULTIVOS", "GlobalCacheConstant.CACHE_CULTIVOS_APP",
+			"GlobalCacheConstant.CACHE_CULTIVO" }, allEntries = true)
 	public void guardar(Object object) {
 		TipoCultivo tipoCultivo = (TipoCultivo) object;
 		tipoCultivoRepository.save(mapearAEntidad(tipoCultivo));
 	}
-	
+
 	@Transactional
-	@CacheEvict(value= {"GlobalCacheConstant.CACHE_CULTIVOS",
-			"GlobalCacheConstant.CACHE_CULTIVOS_APP", 
-			"GlobalCacheConstant.CACHE_CULTIVO"}, allEntries=true)
+	@CacheEvict(value = { "GlobalCacheConstant.CACHE_CULTIVOS", "GlobalCacheConstant.CACHE_CULTIVOS_APP",
+			"GlobalCacheConstant.CACHE_CULTIVO" }, allEntries = true)
 	public void eliminarTipoCultivo(long codigoCultivoSembrado) {
+		List<RequerimientoEntidad> requerimientos = requerimientoRepository
+				.findByCodigoTipoCultivo(codigoCultivoSembrado);
+		if (!requerimientos.isEmpty())
+			throw new ExcepcionClaveForanea(MensajesConstantes.ERROR_ASOCIACION_REQUERIMIENTO);
+
+		List<ParcelaEntidad> parcelas = parcelaRepository.findByCodigoCultivoSembrado(codigoCultivoSembrado);
+		if (!parcelas.isEmpty())
+			throw new ExcepcionClaveForanea(MensajesConstantes.ERROR_ASOCIACION_PARCELA);
+
 		tipoCultivoRepository.deleteById(codigoCultivoSembrado);
 	}
 }
